requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# VLA model support
transformers>=4.35.0
accelerate>=0.24.0
pillow>=10.0.0

# Data handling
h5py>=3.10.0
pyyaml>=6.0

# Logging and tracking (optional)
wandb>=0.16.0
tensorboard>=2.15.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# Note: Isaac Lab must be installed separately
# Follow installation instructions in docs/SETUP.md
